+++ python/os/open.py
@@ -1,0 +1,6 @@
+#!/bin/env python
+import os
+
+f = open('a.out','a')
+f.write('Hello\n')
+f.close()